Move listeners to separate class to lighten Main class Having tons of lines in one class is very bad practice and it's really hard to read! Although this way isn't the best in general it's the best option in current code situation.
